Fix Bluetooth battery level toggle
This line needs to be an OR gate as either of these options should just show the normal icon.

If your headphones have support for the level and the toggle is disabled, the first part will return false (since it is not null) while the second part returns true.

It will also bypass the next statement which means the icon shown will be the enabled but disconnected icon.

By making this an OR, we are saying that if the battery level doesn't exist or the toggle is disabled, we want the normal connected icon.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>