Expose fastscroller text for themes
This commit exposes the fastscroller text so that we can fix a small
issue when using a white accent. The fastscroller drawable takes on the
color of the system accent, while the fastscroller text stays white no
matter what. This creates the problem that we're faced with. This commit
fixes that :-)

Change-Id: I91ab0dc09aa3a8ddd7f5eadfb2eb028e0d9addfc